refactor(clones): split precompute into cohesive siblings
Decompose clone graph precompute into scheduling policy, resumable build, persistence, read-path, and test modules behind a curated mod.rs surface.

Keep IndexDatabase entry points and crate-internal helper paths stable. Preserve atomic generation publication, content-anchor resolution, quiet-window policy, and exact DF/postings eligibility behavior.

Fixes #917
